[Toast] rename refactoring
