--- components/Notifier/index.js
@@ -1,3 +1,0 @@
-// @flow
-import Notifier from './Notifier';
-module.exports = Notifier;
