+++ components/Toast/Toast.less
@@ -1,3 +1,6 @@
+@show-transtion: top 0.2s cubic-bezier(0.22, 0.61, 0.36, 1);
+@hide-transition: opacity 0.15s ease-out;
+
 :local {
   .wrapper {
     position: fixed;
@@ -16,6 +19,7 @@
 
     padding: 10px 20px 11px 20px;
 
+    background: lighten(#333, 80%);
     background: rgba(51, 51, 51, 0.8);
     opacity: 1;
     font-size: 14px;
@@ -57,4 +61,24 @@
     }
   }
 
+}
+
+.slide-and-fade {
+  &-appear, &-enter {
+    top: -40px;
+  }
+
+  &-appear-active, &-enter-active {
+    transition: @show-transtion;
+    top: 0px;
+  }
+
+  &-leave {
+    opacity: 1;
+  }
+
+  &-leave-active {
+    opacity: 0.01;
+    transition: @hide-transition;
+  }
 }
