+++ components/Toast/ToastView.js
@@ -1,0 +1,66 @@
+// @flow
+import React, {Component, PropTypes} from 'react';
+import CROSS from '../internal/cross';
+
+import styles from './Toast.less';
+
+type Props = {
+  children?: string,
+  action?: {
+    label: string,
+    handler: () => void
+  },
+  onClose?: () => void
+}
+
+class ToastView extends Component {
+  props: Props;
+
+  static propTypes = {
+    /**
+     * Adds action handling and close icon fot tost
+     */
+    action: PropTypes.shape({
+      label: PropTypes.string.isRequired,
+      handler: PropTypes.func.isRequired,
+    }),
+    /**
+     * Tost content
+     */
+    children: PropTypes.string.isRequired,
+    onClose: PropTypes.func,
+  }
+
+  render() {
+    const {
+      children,
+      action,
+      onClose,
+      ...rest
+    } = this.props;
+
+    const link = action
+      ? <span className={styles.link} onClick={action.handler} data-close>
+          {action.label}
+        </span>
+      : null;
+
+    const close = action
+      ? <span className={styles.close} onClick={onClose} data-action>
+          {CROSS}
+        </span>
+      : null;
+
+    return (
+      <div className={styles.wrapper}>
+        <div className={styles.root} {...rest}>
+          {children}
+          {link}
+          {close}
+        </div>
+      </div>
+    );
+  }
+}
+
+export default ToastView;
