+++ components/Toast/__tests__/Toast-test.js
@@ -1,41 +1,84 @@
 // @flow
+/* global Lookup */
 import {mount} from 'enzyme';
 import React from 'react';
+import ReactDOM from 'react-dom';
 
 import Toast from '../Toast';
 
+jest.useFakeTimers();
+
 describe('Toast', () => {
-
   it('renders', () => {
-    mount(<Toast>Hello</Toast>);
+    mount(<Toast />);
   });
 
-  it('contains right text', () => {
-    const wrapper = mount(<Toast>Hello</Toast>);
-    expect(wrapper.text()).toBe('Hello');
+  it('doesn\'t throw on push', () => {
+    const wrapper = mount(<Toast />);
+    wrapper.instance().push('message');
   });
 
-  it('shows close if action', () => {
-    const wrapper = mount(
-      <Toast action={{label: 'hello', handler: () => {}}}>Hello</Toast>
-    );
-    expect(wrapper.find('.close')).toBeTruthy();
+  it('sets message to state', () => {
+    const wrapper = mount(<Toast />);
+    wrapper.instance().push('message');
+    expect(wrapper.state().notification).toBe('message');
   });
 
-  it('handles action', () => {
+  it('shows right message', () => {
+    const wrapper = mount(<Toast />);
+    wrapper.instance().push('message');
+
+    const toast = wrapper.instance()._toast;
+    expect(toast).toBeTruthy();
+    expect(ReactDOM.findDOMNode(toast).innerHTML).toContain('message');
+  });
+
+  it('hides message after interval', () => {
+    const wrapper = mount(<Toast />);
+    wrapper.instance().push('message');
+    jest.runAllTimers();
+    const toast = wrapper.instance()._toast;
+    expect(toast).toBeFalsy();
+  });
+
+  it('calls onPush at push', () => {
+    const onPush = jest.fn();
+    const wrapper = mount(<Toast onPush={onPush} />);
+    wrapper.instance().push('somemessage');
+    jest.runAllTimers();
+
+    expect(onPush.mock.calls[0][0]).toBe('somemessage');
+    expect(onPush.mock.calls.length).toBe(1);
+  });
+
+  it('calls onClose after close', () => {
+    const onClose = jest.fn();
+    const wrapper = mount(<Toast onClose={onClose} />);
+    wrapper.instance().push('message');
+    jest.runAllTimers();
+
+    expect(onClose.mock.calls[0][0]).toBe('message');
+    expect(onClose.mock.calls.length).toBe(1);
+  });
+
+  it('support actions in tosts', () => {
+    const wrapper = mount(<Toast />);
+    wrapper.instance().push('message', {label: 'action', handler: () => {}});
+
+    const toast = wrapper.instance()._toast;
+    const link = ReactDOM.
+      findDOMNode(toast).
+      querySelector('[data-action]');
+    expect(link).toBeTruthy();
+  });
+
+  it('passes right actions in tosts', () => {
+    const wrapper = mount(<Toast />);
     const handler = jest.fn();
-    const wrapper = mount(
-      <Toast action={{label: 'hello', handler}}>Hello</Toast>
-    );
-    wrapper.find('.link').simulate('click');
-    expect(handler.mock.calls.length).toBe(1);
-  });
+    wrapper.instance().push('message', {label: 'action', handler});
 
-  it('shows right action label', () => {
-    const wrapper = mount(
-      <Toast action={{label: 'hello', handler: () => {}}}>Hello</Toast>
-    );
-    expect(wrapper.find('.link').text()).toBe('hello');
-  });
+    const toast = wrapper.instance()._toast;
 
+    expect(toast.props.action).toEqual({label: 'action', handler});
+  });
 });
