+++ docs/src/components.js
@@ -126,12 +126,6 @@ module.exports = [
     src: require('!raw!./snippets/Modal-snip.js'),
   },
   {
-    name: 'Notifier',
-    component: require('ui/Notifier'),
-    info: require('!docgen!ui/Notifier/Notifier.js'),
-    src: require('!raw!./snippets/Notifier-snip.js'),
-  },
-  {
     name: 'Radio',
     component: require('ui/Radio'),
     info: require('!docgen!ui/Radio/Radio.js'),
